requirements.txt
requests
schedule
pyyaml
python-dotenv
psycopg2-binary==2.9.9
feedparser
beautifulsoup4
html2text
python-dateutil
openai>=1.40.0
numpy>=1.24
google-api-python-client>=2.137.0
google-auth>=2.33.0
google-auth-httplib2>=0.2.0

